Refactor: Improve type inference for EnumStore
This commit refactors the `EnumStoreType` and `EnumStoreExtension` to improve type inference when getting and setting values.

Previously, the preference key had to be explicitly provided. Now, the type can be inferred from the default value or the value being set.

This change makes the API more concise and easier to use.

Specifically, the following changes were made:

- Added a companion object to `EnumStoreType` with a `getTypeFromValue` function to infer the `EnumStoreType` from a given value.
- Modified the `get` and `set` extension functions in `EnumStoreExtension` to utilize the new `getTypeFromValue` function, removing the need to explicitly pass the `Preferences.Key`.
- Removed the standalone `getPreferenceKey` function from `EnumStoreExtension` as it's now handled within `EnumStoreType`.