Adding `PartOf` annotation to EnumStoreCollection
This commit introduces a new annotation, `@PartOf`, which can be applied to classes implementing `EnumStoreShared`. This annotation allows specifying a parent `EnumStoreCollection` for the shared enum.

The `from` function in `EnumStore` has been updated to use this annotation to determine the collection name for shared enums. If the `@PartOf` annotation is present, its `collection` property will be used; otherwise, the simple name of the shared enum class will be used.

Additionally, `EnumStoreCollection` and `EnumStoreShared` now inherit from a new marker interface, `EnumStoreMarker`.